backend/requirements.txt
fastapi
uvicorn
python-dotenv
python-multipart
boto3
openai>=1.0.0
pypdf
mangum
httpx